File Addtiton and Initial HTML/CSS
I've decided to work with multiple html and css files for better organization of code so I've added quiz.html and quiz.css. I've also changed the name of the script file to quiz.js because it only applies to the quiz itself and not the homepage. The homepage's html and css has officially be set up.