endQuiz and clock Function
I've managed to come up with the function for the quiz ending and clearing the timer if the user answers the question in time. I've also altered the html to fit these new functions. I ended with a bug from earlier so I renmaed some javascript objects as there were musltiple named the same thing. I've also added the clock function. However, I'm not sure how to set a predetermined amount of time. I can set the question to equal a certain length but I do not know how to set the timer for a certain time and have each question diminsh if wrong.