Fixed quiz.js
I've fixed the previous issues and now the quiz.js works and the questions function. There was spelling errors that inhibited the fucntions from working. The forEach issue is still unknown but after retyping the function it works(???). Now I need to complete the highscore javascript to enter in your score.