style.css and README
I already had the css framework done from before while attempting to do the challenge with only one html and script file so I've added that and spread that out to the other new classes and ids. This was fairly simple as I copied the initial style and altered it for each button or text. I'v also centered every thing and made sure all text is accessible.